+++ src/engine/model/AssimpLoader.cpp
@@ -6,6 +6,7 @@
 #include <assimp/postprocess.h>
 #include <assimp/scene.h>
 #include <filesystem>
+#include <stdexcept>
 #include <vector>
 
 using namespace DirectX;
@@ -30,12 +31,11 @@ Model AssimpLoader::Load(const std::string &path) {
     aiMesh *mesh = scene->mMeshes[0];
 
     std::vector<Vertex> vertices;
-    std::vector<uint16_t> indices;
+    std::vector<uint32_t> indices;
 
     vertices.reserve(mesh->mNumVertices);
 
     for (unsigned int i = 0; i < mesh->mNumVertices; i++) {
-
         XMFLOAT3 pos{mesh->mVertices[i].x, mesh->mVertices[i].y,
                      mesh->mVertices[i].z};
 
@@ -50,29 +50,27 @@ Model AssimpLoader::Load(const std::string &path) {
     }
 
     for (unsigned int i = 0; i < mesh->mNumFaces; i++) {
-
         const aiFace &face = mesh->mFaces[i];
 
         for (unsigned int j = 0; j < face.mNumIndices; j++) {
-            indices.push_back((uint16_t)face.mIndices[j]);
+            indices.push_back(face.mIndices[j]);
         }
     }
 
     uint32_t meshId = meshManager_->CreateMesh(
-        vertices.data(), sizeof(Vertex), (uint32_t)vertices.size(),
-        indices.data(), (uint32_t)indices.size());
+        vertices.data(), sizeof(Vertex), static_cast<uint32_t>(vertices.size()),
+        indices.data(), static_cast<uint32_t>(indices.size()));
 
     uint32_t textureId = 0;
 
     if (scene->HasMaterials()) {
-
         aiMaterial *mat = scene->mMaterials[mesh->mMaterialIndex];
 
         aiString texPath;
 
         if (mat->GetTexture(aiTextureType_DIFFUSE, 0, &texPath) == AI_SUCCESS) {
-
             std::filesystem::path modelPath(path);
+
             auto fullPath = modelPath.parent_path() / texPath.C_Str();
 
             textureId = textureManager_->Load(fullPath.wstring());
